day11 finished platform independant code implemented.
