+++ .gitignore
@@ -47,3 +47,4 @@
 
 win32_handmade/Debug/
 build/
+/packages
