+++ win32_handmade/win32_handmade.vcxproj.filters
@@ -18,9 +18,15 @@
     <ClInclude Include="win32_handmade.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="handmade.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="win32_handmade.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="handmade.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
